third_party/WebKit/LayoutTests/fast/borders/border-image-width-numbers-computed-style.html

<!doctype html>
<html>
<head>
  <title>Test that numbers in border-image-{width,outset} are returned correctly</title>
  <script src="../../resources/js-test.js"></script>
  <style>
    .test {
        border-image-width: 0 5.5 4.4 3.3;
        border-image-outset: 2.5 0 1.4 1.9;
        border-style: solid;
        visibility: hidden;
    }
  </style>
</head>
<body>
  <div class="test"></div>
  <script>
    var elm = getComputedStyle(document.querySelector('.test'));
    shouldBeEqualToString("elm.borderImageWidth", "0 5.5 4.4 3.3");
    shouldBeEqualToString("elm.borderImageOutset", "2.5 0 1.4 1.9");
  </script>
</body>
</html>
